+++ readme/RELEASE_NOTES.md
@@ -1,4 +1,4 @@
-* TBD
+* [#5](https://github.com/skuzzle/snapshot-tests/issues/5): Improve support for parameterized tests (`@ParameterizedTest`)
 
 Maven Central coordinates for this release: